fix(mcp): preserve OAuth callback URL context
